pyproject.toml
[project]
authors = [{name = "oomics Team", email = ""}]
name = "rnalens"
requires-python = "==3.9.*"
version = "0.1.0"
dependencies = [ 
    "torch>=2.6.0,<3", 
    "torchvision>=0.21.0,<0.22", 
    "torchsummary>=1.5.1,<2", 
    "tqdm>=4.67.1,<5", 
    "scikit-learn>=1.6.1,<2", 
    "scipy>=1.13.1,<2", 
    "matplotlib>=3.9.4,<4", 
    "seaborn>=0.13.2,<0.14", 
    "transformers[torch]>=4.48.0,<5",
    "einops>=0.6.1,<0.7",
    "biopython>=1.85,<2",
    "streamlit==1.44,<2",
    "liger-kernel>=0.5.6,<0.6"
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[system-requirements]
cuda = ">=11.8"

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
rnalens = { path = ".", editable = true }
